+++ src/collections/resources/faq/layer5-faq.mdx
@@ -1,7 +1,7 @@
 ---
 title: "Layer5 FAQs"
-thumbnail: ../../../assets/images/layer5/layer5-tagline/png/layer5-tag-dark-bg.png
-darkthumbnail: ../../../assets/images/layer5/layer5-tagline/png/layer5-tag-dark-bg.png
+thumbnail: ../../../assets/images/layer5/layer5-only/png/layer5-gradient.png
+darkthumbnail: ../../../assets/images/layer5/layer5-only/png/layer5-gradient.png
 date: 2021-11-02 10:30:05 -0530
 category: FAQ
 tags:
