+++ src/collections/resources/faq/meshery-faq.mdx
@@ -1,7 +1,7 @@
 ---
 title: "Meshery FAQs"
-thumbnail: ../../../assets/images/layer5/layer5-tagline/png/layer5-tag-dark-bg.png
-darkthumbnail: ../../../assets/images/layer5/layer5-tagline/png/layer5-tag-dark-bg.png
+thumbnail: ../../../assets/images/meshery/text-only/meshery-logo-dark-text.png
+darkthumbnail: ../../../assets/images/meshery/text-only/meshery-logo-dark-text.png
 date: 2021-11-03 10:30:05 -0530
 category: FAQ
 tags:
