+++ src/collections/resources/interview/techstrong-tv.mdx
@@ -1,0 +1,27 @@
+---
+title: "TechStrong TV Interview"
+subtitle: "Open Source Summit NA 2022"
+thumbnail: ./techstrong.png
+darkthumbnail: ./techstrong.png
+date: 2022-06-25 10:30:05 -0530
+category: Interview
+tags:
+  - Interview 
+type: Interview
+published: true
+resource: true
+---
+
+
+import { ResourcesWrapper } from "../Resources.style.js";
+import Button from "../../../reusecore/Button";
+import { Link } from "gatsby";
+
+<ResourcesWrapper>
+<p>
+TechStrong TV hosts a variety of live conversations and panel discussions with world’s leading technology experts and leaders at global tech events and user conferences. In this episode of TechStrong TV, straight out of <Link to="/community/events/open-source-summit-north-america-2022">Open Source Summit NA 2022</Link> , catch guest <Link to="https://layer5.io/community/members/lee-calcote">Lee Calcote</Link> from Layer5 and host Alan Shimel discuss the power of <Link to="/projects">Layer5 projects</Link> in managing service meshes, Kubernetes and the rest of your cloud native infrastucture. They also dive into some of the other network-centric CNCF projects like CoreDNS and gRPC. Tune in now! 
+</p>
+ <Button primary url="https://digitalanarchist.com/videos/open-source-summit-na-2022/lee-calcote-layer5" class="btn-center" >
+ <h3>Check out the TechStrong TV Interview with Layer5!</h3>
+ </Button>
+</ResourcesWrapper>
