+++ src/sections/Community/Event-single/index.js
@@ -1,15 +1,13 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Link, graphql, useStaticQuery } from "gatsby";
 import { MDXRenderer } from "gatsby-plugin-mdx";
-import axios from "axios";
 import slugify from "../../../utils/slugify";
-import { Container, Row, Col } from "../../../reusecore/Layout";
+import { Container } from "../../../reusecore/Layout";
 import PageHeader from "../../../reusecore/PageHeader";
-import { Field, Formik, Form } from "formik";
 import EventPageWrapper from "./EventSingle.style";
 import Button from "../../../reusecore/Button";
 import { CTA_Bottom } from "../../../components/Call-To-Actions/CTA_Bottom";
-import layer5_img from "../../../assets/images/layer5/layer5-only/svg/layer5-white-no-trim.svg";
+import CommonForm from "../../../components/CommonForm";
 
 const checkSpeaker = (speaker) => {
 
@@ -55,32 +53,6 @@ const EventSingle = ({ data }) => {
   //const frontmatter = ({speakers = []});
   const { frontmatter, body } = data.mdx;
 
-  const [stepNumber, setStepNumber] = useState(0);
-  // Form values
-  const [memberFormOne, setMemberFormOne] = useState({});
-
-  const [submit, setSubmit] = useState(false);
-
-  const [email, setEmail] = useState("");
-  const [firstname, setFirstName] = useState("");
-  const [lastname, setLastName] = useState("");
-  const [org, setOrg] = useState("");
-  const [occupation, setOccupation] = useState("");
-
-  const nextStep = () => {
-    if (stepNumber === 0) {
-      setSubmit(true);
-    }
-    window.scrollTo(0, 800);
-  };
-
-  useEffect(() => {
-    if (submit) {
-      axios.post("https://hook.us1.make.com/gguommoyd14634ur9xs7l37widuoa7e9", {
-        memberFormOne,
-      });
-    }
-  }, [submit]);
 
   return (
     <EventPageWrapper>
@@ -93,55 +65,14 @@ const EventSingle = ({ data }) => {
           <div className="event-info-block">
             <div className="tags">
 
-              { frontmatter.register && stepNumber === 0 &&
-                <div className="form-body">
-                  <h3 className="form-title">Register Today!</h3>
-                  <Formik
-                    initialValues={{
-                      firstname: firstname,
-                      lastname: lastname,
-                      email: email,
-                      occupation: occupation,
-                      org: org,
-                      form: "event",
-                    }}
-                    onSubmit={values => {
-                      setMemberFormOne(values);
-                      setStepNumber(1);
-                      nextStep();
-                      setFirstName(values.firstname);
-                      setEmail(values.email);
-                      setLastName(values.lastname);
-                      setOccupation(values.occupation);
-                      setOrg(values.org);
-                    }}
-                  >
-                    <Form className="form" method="post">
-                      <Row>
-                        <Col xs={6}>
-                          <label htmlFor="firstname" className="form-name">First Name <span className="required-sign">*</span></label>
-                          <Field type="text" className="text-field" id="firstname" name="firstname" maxLength="32" pattern="[A-Za-z]{1,32}" required />
-                        </Col>
-                        <Col xs={6}>
-                          <label htmlFor="lastname" className="form-name">Last Name <span className="required-sign">*</span></label>
-                          <Field type="text" className="text-field" id="lastname" name="lastname" maxLength="32" pattern="[A-Za-z]{1,32}" required />
-                        </Col>
-                      </Row>
-                      <label htmlFor="email" className="form-name">Email Address <span className="required-sign">*</span></label>
-                      <Field type="text" className="text-field" id="email" name="email" pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$" required />
-                      <label htmlFor="occupation" className="form-name">Occupation / Title <span className="required-sign">*</span></label>
-                      <Field type="text" className="text-field" id="occupation" name="occupation" required />
-                      <label htmlFor="org" className="form-name">Organization / Company / School <span className="required-sign">*</span></label>
-                      <Field type="text" className="text-field" id="org" name="org" />
-
-                      <Button secondary type="submit" className="btn" title="Submit" />
-                    </Form>
-                  </Formik>
-                </div>
-              }
-              {
-                stepNumber === 1 &&
-                    <ThankYou title={frontmatter.title} />
+              { frontmatter.register &&
+                  <CommonForm
+                    title="Register Today!"
+                    form="event"
+                    account_desc="Please provide at least one account."
+                    submit_title = {`Thank you for registering to ${frontmatter.title}!`}
+                    submit_body = {`You are now signed up for the ${frontmatter.title} workshop by Layer5. Please patiently await your acceptance. We'll send out additional information about the event soon.`}
+                  />
               }
 
               <MDXRenderer>{body}</MDXRenderer>
@@ -175,19 +106,6 @@ const EventSingle = ({ data }) => {
         </Container>
       </div>
     </EventPageWrapper>
-  );
-};
-
-const ThankYou = (props) => {
-  return (
-    <Container>
-      <div className="thank-you-box">
-        <h2>Thank you for registering to {props.title}!</h2>
-        <p>You are now signed up for the <i>{props.title}</i> workshop by Layer5. Please patiently await your acceptance and start of the beta program. We are working through a growing waitlist </p>
-        <p>In the meantime, please visit our <a href="https://discuss.layer5.io">community forum</a> or join us in our <a href="https://slack.layer5.io">community Slack</a>.</p>
-        <h3 className="white">- Team <img src={layer5_img} alt="Layer5" width="125" /></h3>
-      </div>
-    </Container>
   );
 };
